+++ src/main/java/com/goorm/friendchise/domain/headquarter/application/KakaoApiService.java
@@ -15,77 +15,7 @@ import java.util.*;
 @Service
 @RequiredArgsConstructor
 public class KakaoApiService {
-    private final WebClient webClient;
-
-    /*
-        * "키워드로 장소 검색하기" 카카오 지도 API를 사용하여 radius 반경 내 장소 정보를 가져온다. (동기)
-        * @param keyword 검색 키워드
-        * @param y 위도
-        * @param x 경도
-        * @param radius 검색 반경
-        * @return KakaoSearchResDto
-    */
-    public KakaoApiResultDto requestPlaceDataByKeywordSync(String keyword, Double y, Double x, int radius) {
-        String uri = makeKeywordSearchAPIUri(keyword, y, x, radius);
-
-        return webClient.get()
-                .uri(uri)
-                .retrieve()
-                .bodyToMono(KakaoApiResultDto.class)
-                .block();
-    }
-
-    /*
-     * "키워드로 장소 검색하기" 카카오 지도 API를 사용하여 radius 반경 내 장소 정보를 가져온다. (비동기)
-     * @param keyword 검색 키워드
-     * @param y 위도
-     * @param x 경도
-     * @param radius 검색 반경
-     * @return Mono<KakaoSearchResDto>
-     */
-    public Mono<KakaoApiResultDto> requestPlaceDataByKeywordAsync(String keyword, Double y, Double x, int radius) {
-        String uri = makeKeywordSearchAPIUri(keyword, y, x, radius);
-
-        return webClient.get()
-                .uri(uri)
-                .retrieve()
-                .bodyToMono(KakaoApiResultDto.class);
-    }
-
-    /*
-     * "카테고리로 장소 검색하기" 카카오 지도 API를 사용하여 radius 반경 내 장소 정보를 가져온다. (동기)
-     * @param categoryGroupCode 카테고리 그룹 코드
-     * @param y 위도
-     * @param x 경도
-     * @param radius 검색 반경
-     * @return KakaoSearchResDto
-     */
-    public KakaoApiResultDto requestPlaceDataByCategorySync(String categoryGroupCode, Double y, Double x, int radius) {
-        String uri = makeCategorySearchAPIUri(categoryGroupCode, y, x, radius);
-
-        return webClient.get()
-                .uri(uri)
-                .retrieve()
-                .bodyToMono(KakaoApiResultDto.class)
-                .block();
-    }
-
-    /*
-        * "카테고리로 장소 검색하기" 카카오 지도 API를 사용하여 radius 반경 내 장소 정보를 가져온다. (비동기)
-        * @param categoryGroupCode 카테고리 그룹 코드
-        * @param y 위도
-        * @param x 경도
-        * @param radius 검색 반경
-        * @return Mono<KakaoSearchResDto>
-     */
-    public Mono<KakaoApiResultDto> requestPlaceDataByCategoryAsync(String categoryGroupCode, Double y, Double x, int radius) {
-        String uri = makeCategorySearchAPIUri(categoryGroupCode, y, x, radius);
-
-        return webClient.get()
-                .uri(uri)
-                .retrieve()
-                .bodyToMono(KakaoApiResultDto.class);
-    }
+    private final KakaoApiClient kakaoApiClient;
 
     public Mono<Map<String, KakaoApiResultDto>> getTotalPlaceData(
             String franchiseName,
@@ -95,10 +25,8 @@ public class KakaoApiService {
             Double y,
             Double x
     ) {
-        // TODO: 각 api 호출 프로세스를 메소드로 분리하는 리팩토링
-
         // 1. 동일 프랜차이즈 매장 검색
-        KakaoApiResultDto sameFranchiseStoreResult = requestPlaceDataByKeywordSync(franchiseName, y, x, 500);
+        KakaoApiResultDto sameFranchiseStoreResult = kakaoApiClient.requestPlaceDataByKeywordSync(franchiseName, y, x, 500);
         if(!sameFranchiseStoreResult.documents().isEmpty()) {
             return null; // 동일 프랜차이즈 매장이 존재하면 바로 리턴
         }
@@ -108,15 +36,15 @@ public class KakaoApiService {
         // 2. 동일 업종 경쟁 매장 검색
         Mono<KakaoApiResultDto> sameCategoryStoreResult;
         if(subCategory.equals(SubCategory.NONE)) { // subCategory가 없으면 category로 검색
-            sameCategoryStoreResult =  requestPlaceDataByKeywordAsync(category.getValue(), y, x, 1000);
+            sameCategoryStoreResult =  kakaoApiClient.requestPlaceDataByKeywordAsync(category.getValue(), y, x, 1000);
         } else {
-            sameCategoryStoreResult = requestPlaceDataByKeywordAsync(subCategory.getValue(), y, x, 1000);
+            sameCategoryStoreResult = kakaoApiClient.requestPlaceDataByKeywordAsync(subCategory.getValue(), y, x, 1000);
         }
         totalSearchResults.put("반경 1km 내 동일 업종 경쟁 매장", sameCategoryStoreResult);
 
         // 3. 버스 정류장, 지하철역 검색
-        Mono<KakaoApiResultDto> busStopResult = requestPlaceDataByKeywordAsync("버스정류장", y, x, 200);
-        Mono<KakaoApiResultDto> subwayStationResult = requestPlaceDataByCategoryAsync(CategoryGroupCode.SUBWAY.getCode(), y, x, 500);
+        Mono<KakaoApiResultDto> busStopResult = kakaoApiClient.requestPlaceDataByKeywordAsync("버스정류장", y, x, 200);
+        Mono<KakaoApiResultDto> subwayStationResult = kakaoApiClient.requestPlaceDataByCategoryAsync(CategoryGroupCode.SUBWAY.getCode(), y, x, 500);
         totalSearchResults.put("반경 200m 내 버스정류장", busStopResult);
         totalSearchResults.put("반경 500m 내 지하철역", subwayStationResult);
 
@@ -124,7 +52,7 @@ public class KakaoApiService {
         for(String selectedCategory : userSelectedCategory) {
             CategoryGroupCode categoryGroupCode = CategoryGroupCode.fromString(selectedCategory);
             if(categoryGroupCode == null) continue;
-            Mono<KakaoApiResultDto> userDefinedResult = requestPlaceDataByCategoryAsync(categoryGroupCode.getCode(), y, x, 500);
+            Mono<KakaoApiResultDto> userDefinedResult = kakaoApiClient.requestPlaceDataByCategoryAsync(categoryGroupCode.getCode(), y, x, 500);
             totalSearchResults.put("반경 500m 내 " + selectedCategory, userDefinedResult);
         }
 
@@ -141,31 +69,4 @@ public class KakaoApiService {
         });
 
     }
-
-    // 키워드로 장소 검색하기 API URI 생성
-    private String makeKeywordSearchAPIUri(String keyword, Double y, Double x, int radius) {
-        return UriComponentsBuilder.fromPath("/search/keyword.json")
-                .queryParam("query", keyword)
-                .queryParam("x", x)
-                .queryParam("y", y)
-                .queryParam("radius", radius) // 값 조정 필요
-                .queryParam("size", 10)
-                .queryParam("sort", "distance")
-                .build()
-                .toUriString();
-    }
-
-    // 카테고리로 장소 검색하기 API URI 생성
-    private String makeCategorySearchAPIUri(String categoryGroupCode, Double y, Double x, int radius) {
-        return UriComponentsBuilder.fromPath("/search/category.json")
-                .queryParam("category_group_code", categoryGroupCode)
-                .queryParam("x", x)
-                .queryParam("y", y)
-                .queryParam("radius", radius) // 값 조정 필요
-                .queryParam("size", 10)
-                .queryParam("sort", "distance")
-                .build()
-                .toUriString();
-    }
-
 }
